backend/requirements/base.txt



django
djangorestframework
django-cors-headers   # handle CORS headers to connent wiht react
pillow
django-cleanup   # clean images after delete instance
python-decouple  # .env variables handle
djangorestframework-simplejwt
django-unfold    # admin theme
django-import-export  # import and export data